refactor: Entity Package 수정
