+++ src/main/java/com/tmax/commerce/itemmodule/entity/base/PersistableDateTimeEntity.java
@@ -1,4 +1,4 @@
-package com.tmax.commerce.itemmodule.entity;
+package com.tmax.commerce.itemmodule.entity.base;
 
 
 import jakarta.persistence.MappedSuperclass;
