+++ src/main/java/com/tmax/commerce/itemmodule/repository/ItemRepository.java
@@ -1,7 +1,6 @@
 package com.tmax.commerce.itemmodule.repository;
 
-import com.tmax.commerce.itemmodule.entity.Item;
-import org.springframework.data.jpa.repository.JpaRepository;
+import com.tmax.commerce.itemmodule.entity.item.Item;
 
 import java.util.UUID;
